Move the mock-data.js file to the /lib
- for easier access across all component that need this file
- I've also created prop types for each object interface which will help
  developers to use this type wherever the mock data needs to be
  imported
