+++ src/pages/FirstGroupPage/components/LastStep.jsx
@@ -2,7 +2,7 @@ import { BodyText, Heading } from "@/components/Typography"
 import { useFirstGroupPageContext } from "../hooks/useFirstGroupPageContext"
 import { StepActions } from "../FirstGroupPage"
 import { StepsContentCommonTypes } from "./types"
-import { EXPENSES_MOCK_DATA, PARTICIPANTS_MOCK_DATA } from "../mock-data"
+import { EXPENSES_MOCK_DATA, PARTICIPANTS_MOCK_DATA } from "@/lib/mock-data"
 import { ExpensesList } from "@/components/ExpensesList"
 import { GroupInfoWidget } from "@/components/GroupInfoWidget"
 import { SelectedParticipantsList } from "@/components/SelectedParticipantsList"
